display game day results, added teams seeder info

+++ app/Http/Controllers/GameDaysController.php
@@ -12,6 +12,8 @@ use COVL\Season;
 use COVL\Point;
 use COVL\Http\Requests\StoreGameDayRequest;
 
+use Illuminate\Database\Eloquent\Collection;
+
 class GameDaysController extends Controller {
 
 	public function __construct(GameDay $game_day, Season $season, Game $game, GameSet $game_set, Point $point) 
@@ -27,7 +29,7 @@ class GameDaysController extends Controller {
 
 	public function index(Request $request) 
         {
-                $game_days = $this->game_day->where('hidden', '0')->paginate(2);
+                $game_days = $this->game_day->where('hidden', '0')->paginate(20);
 
                 if ($request->ajax()) {
                         $html = view('game_days.list', compact('game_days'))->render();
@@ -99,50 +101,54 @@ class GameDaysController extends Controller {
 		$game_day = $this->game_day->find($id);
                 if ($game_day instanceof GameDay) {
 		
-			foreach ($request->game_score as $game_id => $game_data) {
-				$game = $this->game->find($game_id);
-				if ($game instanceof Game) {
-					foreach ($game_data as $set_number => $set_info) {
-						if (array_key_exists('game_set_id', $set_info)) {
-							$game_set = $this->game_set->find($set_info['game_set_id']);
-						} else {
-							$game_set = new GameSet;
-						}
-						$game_set->game_id = $game->id;
-						$game_set->number = $set_number;
-						$game_set->save();
+			if (count($request->game_score) > 0) {
+				foreach ($request->game_score as $game_id => $game_data) {
+					$game = $this->game->find($game_id);
+					if ($game instanceof Game) {
+						foreach ($game_data as $set_number => $set_info) {
+							if (array_key_exists('game_set_id', $set_info)) {
+								$game_set = $this->game_set->find($set_info['game_set_id']);
+							} else {
+								$game_set = new GameSet;
+							}
+							$game_set->game_id = $game->id;
+							$game_set->number = $set_number;
+							$game_set->save();
 
-						$old_home_count = count($this->point->where('game_set_id', $game_set->id)->where('team_id', $game->home_team->id)->get());
-						$old_away_count = count($this->point->where('game_set_id', $game_set->id)->where('team_id', $game->away_team->id)->get());
-						
-						if ($old_home_count < $set_info['home_team']) {
-							for ($i = $old_home_count; $i < $set_info['home_team']; $i++) {
-								$point = new Point;
-								$point->game_set_id = $game_set->id;
-								$point->team_id = $game->home_team->id;
-								$point->save();
+							$old_home_count = count($this->point->where('game_set_id', $game_set->id)->where('team_id', $game->home_team->id)->get());
+							$old_away_count = count($this->point->where('game_set_id', $game_set->id)->where('team_id', $game->away_team->id)->get());
+							
+							if ($old_home_count < $set_info['home_team']) {
+								for ($i = $old_home_count; $i < $set_info['home_team']; $i++) {
+									$point = new Point;
+									$point->game_set_id = $game_set->id;
+									$point->team_id = $game->home_team->id;
+									$point->save();
+								}
+							} else if($old_home_count > $set_info['home_team'])  {
+								$delete_count = $old_home_count - $set_info['home_team'];
+								$this->point->where('game_set_id', $game_set->id)->where('team_id', $game->home_team->id)->orderBy('id', 'desc')->limit($delete_count)->delete();
 							}
-						} else if($old_home_count > $set_info['home_team'])  {
-							$delete_count = $old_home_count - $set_info['home_team'];
-							$this->point->where('game_set_id', $game_set->id)->where('team_id', $game->home_team->id)->orderBy('id', 'desc')->limit($delete_count)->delete();
-						}
-	
-						if ($old_away_count < $set_info['away_team']) {						
-							for ($i = $old_away_count; $i < $set_info['away_team']; $i++) {
-								$point = new Point;
-								$point->game_set_id = $game_set->id;
-								$point->team_id = $game->away_team->id;
-								$point->save();
+		
+							if ($old_away_count < $set_info['away_team']) {						
+								for ($i = $old_away_count; $i < $set_info['away_team']; $i++) {
+									$point = new Point;
+									$point->game_set_id = $game_set->id;
+									$point->team_id = $game->away_team->id;
+									$point->save();
+								}
+							} else if ($old_away_count > $set_info['away_team']) {
+								$delete_count = $old_away_count - $set_info['away_team'];
+								$this->point->where('game_set_id', $game_set->id)->where('team_id', $game->away_team->id)->orderBy('id', 'desc')->limit($delete_count)->delete();
 							}
-						} else if ($old_away_count > $set_info['away_team']) {
-							$delete_count = $old_away_count - $set_info['away_team'];
-							$this->point->where('game_set_id', $game_set->id)->where('team_id', $game->away_team->id)->orderBy('id', 'desc')->limit($delete_count)->delete();
 						}
 					}
 				}
-			}
 
-                        return redirect()->route('game-days.games', ['id' => $game_day->id]);
+				return redirect()->route('game-days.games', ['id' => $game_day->id]);
+			} else {
+				return redirect()->back();
+			}
                 } else {
                         return redirect()->back();
                 }
@@ -164,9 +170,7 @@ class GameDaysController extends Controller {
 	public function results($id) {
 		$game_day = $this->game_day->find($id);
                 if ($game_day instanceof GameDay) {
-
 			return view('game_days.results', compact('game_day'));
-
                 } else {
                         return redirect()->back();
                 }
